chore: test peerDep and workflow (#267)
Adds the ability to use an arbitrary version of KFC by listing it as a
`peerDependency` and installing as a dependency in the top level
`package.json`. This adds tests to ensure that everything still works.

There was a docker inspect command that was trying to inspect `pepr:dev`
which was the default value for image. Docker has not loaded in any
image at that point so the command would have failed unless you had
loaded pepr:dev into the container. This test is a generic test to
ensure that the local changes had not affected the e2e tests, so the
goal is to test based on the version in the package.json, which is
pulled from GHCR upon deployment of the module.

Pepr explicitly passes in the pepr:dev image while executing e2e tests
https://github.com/defenseunicorns/pepr/blob/main/.github/workflows/pepr-excellent-examples.yml#L202

Closes
https://github.com/defenseunicorns/pepr-excellent-examples/issues/231
Closes
https://github.com/defenseunicorns/pepr-excellent-examples/issues/266

---------

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>