chore: increase moduleUp timelimit (#295)
We have been hitting timeouts in the runners. It seems more frequently
right now that the whole module is not coming up as the webhook is not
reachable by Kubernetes. If we increase the initial limit around
actually deploying the webhook those pods are more likely to be up and
healthy and the theory is we will get less flakes.

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>