chore: package.json custom and default are using outdated versions (#306)
After doing some research on the custom rbac test, i realized that this
test is using 3 different `package.json` files during the test. The
problem, however is that the versions of Pepr in the non default
package.json have not been updated to reflect the version of Pepr that
the rest of the tests are using.

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>