chore: update pepr installation and check (#313)
fixes #312 

Since switching to ESLint v9.x we had to update dependencies in Pepr. We
also had to change logic in PEXEX to install the nightly version of pepr
as seen in this
[PR](https://github.com/defenseunicorns/pepr-excellent-examples/pull/307/files).
Before we were installing npm i pepr, and seeing check if the pepr env
is pepr. Now we are installing pepr@nightly which makes this check never
return positive getPeprAlias() !== 'pepr'. We need to change that to
.StartsWith("pepr")` so we can return true for a nightly and also we
need to force install the dependencies since there are some clashes due
to newer versions of dependencies in the nightly.

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>