chore: configure husky hooks for project (#211)
Addresses the configuration needed for #186. Relates to #189.

This PR uses an eslint ruleset shared by Pepr, but every single finding
is set to `warn`. The intent of this PR is to create a mechanism to
provide eslint feedback to developers on every commit. Enforcing or
resolving linter findings for PEXEX is left for separate PRs.