chore: uuid test in hello-pepr-mutate (#335)
We are in the process of removing our journey tests in Pepr in exchange
for PEXEX which provides more comprehensive coverage. However, we did
not have a PEXEX (pepr-excellent-examples) test that covered `npx pepr
uuid` as we do in the journeys. This PR adds a test case for uuid.

Fixes #289

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>