chore: npx pepr monitor should display validation results to console (#334)
We are in the process of removing our journey tests in Pepr in exchange
for PEXEX which provides more comprehensive coverage. However, we did
not have a PEXEX (pepr-excellent-examples) test that covered `npx pepr
monitor` as we do in the journeys. This PR adds a test case for monitor.


Fixes #290

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>