chore: module up should look for controller startup complete to avoid conflicts with controllers (#337)
PR #[2359](https://github.com/defenseunicorns/pepr/pull/2359) in Pepr
updated the build and deploy to only deploy necessary resources for the
current module.

This means if you module does not use watch, then the watch manifests
will not be _deployed_ or _build_, and likewise with `Mutate()` and
`Validate()`.

The problem with moduleUp is that it is assuming the module is ready
when `Scheduling processes` is logged twice. This will not be true after
the PR is merged. If it is a watch processor then it will only be logged
once. If it is a admission processor then it only needs to be logged
once (1 pod will be in a ready state - the other replica that is not
passing readiness probes will not serve traffic), if it is watch and
admission then we need it logged 3 times since we will not be able to
distinguish where the message came from, this is the only way to ensure
the watcher is ready.

---------

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>